Fixed UI in saving and logging out in the app.
